authentication/requirements.txt
fastapi==0.104.1
fastapi-users[beanie]==12.1.2
fastapi-users[oauth]==12.1.2 
beanie==1.23.6
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.2.1
httpx==0.25.2
pydantic-settings==2.0.3
boto3==1.34.0
python-dotenv==1.0.0
uvicorn==0.24.0
motor==3.3.1
pymongo==4.3.3
bcrypt==3.2.2
